+++ src/main/java/com/raoulvdberge/refinedpipes/RefinedPipesItems.java
@@ -7,6 +7,10 @@ import net.minecraftforge.registries.ObjectHolder;
 public class RefinedPipesItems {
     @ObjectHolder(RefinedPipes.ID + ":basic_item_pipe")
     public static final BaseBlockItem BASIC_ITEM_PIPE = null;
-    @ObjectHolder(RefinedPipes.ID + ":extractor_attachment")
-    public static final AttachmentItem EXTRACTOR_ATTACHMENT = null;
+    @ObjectHolder(RefinedPipes.ID + ":basic_extractor_attachment")
+    public static final AttachmentItem BASIC_EXTRACTOR_ATTACHMENT = null;
+    @ObjectHolder(RefinedPipes.ID + ":improved_extractor_attachment")
+    public static final AttachmentItem IMPROVED_EXTRACTOR_ATTACHMENT = null;
+    @ObjectHolder(RefinedPipes.ID + ":advanced_extractor_attachment")
+    public static final AttachmentItem ADVANCED_EXTRACTOR_ATTACHMENT = null;
 }