+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Refined Pipes Changelog
 
 ## 0.1.1
-- Renamed Pipes to Item Pipe (raoulvdberge)
+- Renamed "Pipe" to "Item Pipe" (raoulvdberge)
+- Added different tiers for the Extractor Attachment: Basic, Improved and Advanced (raoulvdberge)
 
 ## 0.1
 - Initial release